[#186] refactor(Post): TutorRes 패키지 변경
